fix: split layout creation for each function
